requirements.txt
python-dotenv==1.0.1
httpx==0.27.0
websockets<12.0
solana==0.34.2
solders==0.21.0
borsh-construct==0.1.0